feat: add `debug` option
When true, the "/debug/collect" endpoint is used, which responds with validation errors.

NOTE: Hits are never recorded while in debug mode.

The `baseURL` option takes precedence over the `debug` option, so you need to provide a different `baseURL` in addition to `debug: true` if you want to debug requests to a custom url.